+++ src/qcommon/vm_local.h
@@ -25,6 +25,9 @@ Foundation, Inc., 51 Franklin St, Fifth Floor, Boston, MA  02110-1301  USA
 #include "q_shared.h"
 #include "qcommon.h"
 
+//use alternative float casting behavior, affects physics behavior
+#define VM_ALT_FLOAT_CASTING
+
 #define	MAX_OPSTACK_SIZE 512
 #define	PROC_OPSTACK_SIZE 30